requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi
uvicorn[standard]

# ── LLM ───────────────────────────────────────────────────────────────────────
anthropic                     # Claude API (active)
openai                        # Azure OpenAI — kept for reference; no longer called

# ── Data Sources ──────────────────────────────────────────────────────────────
polygon-api-client
fredapi
yfinance
requests
beautifulsoup4                # Fed speech scraping

# ── Database ──────────────────────────────────────────────────────────────────
supabase
chromadb                      # Local vector store for research memos

# ── Quant / Data ──────────────────────────────────────────────────────────────
pandas
numpy
scipy

# ── Scheduling ────────────────────────────────────────────────────────────────
apscheduler

# ── Broker ────────────────────────────────────────────────────────────────────
ib_insync

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv
pydantic
sentence-transformers             # local embeddings (BAAI/bge-base-en-v1.5); ~400MB model download on first run
